rbedrock: Analysis and Manipulation of Data from Minecraft Bedrock Edition

Implements an interface to Minecraft (Bedrock Edition) worlds. Supports the analysis and management of these worlds and game saves.

Details

rbedrock is currently under development.

See also

Author

Maintainer: Reed Cartwright racartwright@gmail.com (ORCID)

Other contributors:

Rich FitzJohn rich.fitzjohn@gmail.com [contributor]

Christian Stigen Larsen [contributor]

The LevelDB Authors [copyright holder]